Implement a struct representing the Game Boy
Make this struct a container for the various components of the Game
Boy (currently only the CPU and memory). As part of this, decouple the
CPU and memory by moving the Game Boy's memory from the CPU struct to
the Game Boy struct. Refactor the CPU instruction functions to use the
Game Boy struct.