Refactor makefile rules for making directories and building
Rewrite the way that the required directories are created so that
`mkdir -p` is not called unnecessarily (i.e. only called when the
required directories don't exist). In addition, refactor the building
of the emulator so that all the object files are stored in a single
directory, as opposed to mirroring the directory structure of the
source files.