Implement the timer and divider registers
More specifically, implement the DIV register, the Game Boy's internal
clock counter (of which DIV is the upper byte), and the TIMA, TMA, and
TAC registers. Additionally, implement a function that will increment
DIV, TIMA, and the internal clock counter as the CPU clock ticks.
Lastly, implement the incrementing behavior of TIMA when certain
conditions are met during a write to the DIV or TAC registers.